Refactored to conform to existing architecture. Continue refactoring by moving logic from the viewModel to the data and domain layers
