+++ godtools/App/Features/Dashboard/Presentation/Dashboard/DashboardView.swift
@@ -24,93 +24,6 @@ struct DashboardView: View {
         self.viewModel = viewModel
     }
 
-    func checkNotificationStatus() async -> String {
-        let center = UNUserNotificationCenter.current()
-        let settings = await center.notificationSettings()
-
-        switch settings.authorizationStatus {
-        case .authorized, .provisional, .ephemeral:
-            print("NotificationStatus: Enabled")
-            return "Enabled"
-        case .denied:
-            print("NotificationStatus: Denied")
-            return
-                "Denied"
-        case .notDetermined:
-            print("NotificationStatus: Undetermined")
-            return "Undetermined"
-        @unknown default:
-            return "Unknown"
-        }
-    }
-
-    func recordLastPrompt() {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "MM/dd/yyyy"
-
-        guard let testDate = dateFormatter.date(from: "01/01/2025") else {
-            print("Failed to create date from string.")
-            return
-        }
-
-        let dateString = dateFormatter.string(from: testDate)
-        UserDefaults.standard.set(dateString, forKey: "lastPrompted")
-        print("set lastPrompted to: \(dateString)")
-    }
-
-    func shouldPromptNotificationsSheet() {
-        Task {
-            let notificationStatus = await checkNotificationStatus()
-
-            // Retrieve lastPrompted date
-            let dateFormatter = DateFormatter()
-            dateFormatter.dateFormat = "MM/dd/yyyy"
-            let lastPrompted = UserDefaults.standard.string(
-                forKey: "lastPrompted")
-            let lastPromptedDate =
-                lastPrompted.flatMap {
-                    dateFormatter.date(from: $0)
-                } ?? Date.distantPast
-
-            // Get current date
-            let currentDate = Date()
-            let calendar = Calendar.current
-            let twoMonthsAgo = calendar.date(
-                byAdding: .month, value: -2, to: currentDate)!
-
-            if notificationStatus != "Approved" {
-                if notificationStatus == "Undetermined" {
-                    print("Detected undetermined status")
-                    viewModel.getNotificationsViewModel().bottomSheetPosition =
-                        .dynamicTop
-                    recordLastPrompt()
-                } else if notificationStatus == "Denied"
-                    && lastPromptedDate < twoMonthsAgo
-                {
-                    print(
-                        "Previously denied but last prompt was more than two months ago."
-                    )
-                    print(
-                        "Bottom sheet position before: \(viewModel.getNotificationsViewModel().bottomSheetPosition)"
-                    )
-                    viewModel.getNotificationsViewModel().bottomSheetPosition =
-                        .dynamicTop
-                    print(
-                        "Bottom sheet position after : \(viewModel.getNotificationsViewModel().bottomSheetPosition)"
-                    )
-                    recordLastPrompt()
-                } else if notificationStatus == "Denied"
-                    && lastPromptedDate > twoMonthsAgo
-                {
-                    print(
-                        "Previously denied and prompted recently.")
-                    recordLastPrompt()
-
-                }
-            }  // already approved
-        }
-    }
-
     var body: some View {
 
         GeometryReader { geometry in
@@ -118,7 +31,11 @@ struct DashboardView: View {
             VStack(alignment: .center, spacing: 0) {
 
                 Button(action: {
-                    shouldPromptNotificationsSheet()
+                    Task {
+                        await viewModel.getNotificationsViewModel()
+                            .shouldPromptNotificationsSheet()
+                    }
+
                 }) {
 
                     Text("shouldPromptNotificationsSheet()")
