+++ godtools/App/Features/OptInNotification/Domain/Entities/OptInNotificationsInterfaceStringsDomainModel.swift
@@ -1,5 +1,5 @@
 //
-//  OptInForNotificationsStringsDomainModel.swift
+//  OptInForNotificationsInterfaceStringsDomainModel.swift
 //  godtools
 //
 //  Created by Levi Eggert on 3/10/25.
@@ -12,6 +12,6 @@ struct OptInNotificationsStringsDomainModel {
     
     let title: String
     let body: String
-    let enableNotificationsActionTitle: String
-    let enableLaterActionTitle: String
+    let allowNotificationsActionTitle: String
+    let maybeLaterActionTitle: String
 }
