+++ godtools/App/Features/OptInNotification/Domain/Interface/GetOptInNotificationsInterfaceStringsRepositoryInterface.swift
@@ -1,0 +1,16 @@
+//
+//  GetNotificationsInterfaceStringsRepository.swift
+//  godtools
+//
+//  Created by Jason Bennett on 3/13/25.
+//  Copyright © 2025 Cru. All rights reserved.
+//
+
+
+import Foundation
+import Combine
+
+protocol GetOptInNotificationsInterfaceStringsRepositoryInterface {
+    
+    func getStringsPublisher(translateInLanguage: AppLanguageDomainModel) -> AnyPublisher<OptInNotificationsStringsDomainModel, Never>
+}
